fonctionne en float (a tester en vraix)

+++ Comparaison-image/Comparaison-image.h
@@ -2,11 +2,12 @@
 #define COMPARAISON_IMAGE_C
 #include <stdio.h>
 #include <stdlib.h>
+#include <math.h>
 
 typedef struct
 {
     int id;
-    int pourcentage;
+    float pourcentage;
 } tab_similaire;
 
 tab_similaire* Comparaison_descripteur_image(int *Erreur, char PathRecueil[], char PathNvImg[], int Nb_Bits_Fort, int *Nb_des);
